feat:added test VC test cases - Verify a JWT signature is valid - When  a JWT signature is invalid - Separated the create-credential and share-credential tests #142
